refactor: Validation 열거형을 TextField 내부로 이동
